+++ doc/CHANGELOG.md
@@ -1,10 +1,14 @@
 # What's New with Enterprise Web Components
 
+## 1.0.0-beta.9
+
+### 1.0.0-beta.9 Fixes
+
 ## 1.0.0-beta.8
 
 ### 1.0.0-beta.8 Fixes
 
-- `[Button/Input/Dropdown]` Added posibility to show loading indicator. ([#1048](https://github.com/infor-design/enterprise-wc/issues/1048))
+- `[Button/Input/Dropdown]` Added ability to show loading indicator. ([#1048](https://github.com/infor-design/enterprise-wc/issues/1048))
 - `[Datagrid]` Added feature to export data grid to excel xlsx or csv file. ([#153](https://github.com/infor-design/enterprise-wc/issues/153))
 - `[Datagrid]` Fixed event-handling bug where `contextmenu` event listeners were never cleaned up, causing multiple selection events to occur. ([#1174](https://github.com/infor-design/enterprise-wc/issues/1174))
 - `[Datagrid/Dropdown]` Separated IdsDropdownList into its own component, re-integrated the new component into IdsDropdown, and fixed containment/cutoff issues in IdsDataGrid using the new list component. ([#1065](https://github.com/infor-design/enterprise-wc/issues/1065))